refactor(spx-gui):Rewrite export function into useSvgExporter and decoupling
